api_yamdb/api/v1/mixins.py
from rest_framework import filters, mixins, viewsets
from rest_framework.pagination import LimitOffsetPagination
from .permissions import IsAdmin


class SlugNameViewSet(
    mixins.CreateModelMixin,
    mixins.ListModelMixin,
    mixins.DestroyModelMixin,
    viewsets.GenericViewSet
):
    permission_classes = (IsAdmin, )
    pagination_class = LimitOffsetPagination
    lookup_field = 'slug'
    filter_backends = (filters.SearchFilter,)
    search_fields = ('=slug', 'name')
